+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Adding filter and parties to `OrderMassCancelRequest` and `OrderMassCancelReport` (#414)
+
 ## 0.9.7 &ndash; 2023-09-18
 
 ## 0.9.6 &ndash; 2023-07-22